LayoutTests/inspector/dom-debugger/dom-breakpoint-node-removed-direct-expected.txt

CONSOLE MESSAGE: BREAKPOINT ACTION LOG 1
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 2
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 3
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 4
Tests for DOM breakpoints.


== Running test suite: DOMBreakpoint.NodeRemoved
-- Running test case: DOMBreakpoint.NodeRemoved.Direct.BreakpointEnabled
Adding "node-removed:1,HTML,1,BODY,1,DIV,0,DIV" DOM Breakpoint...
PAUSED:
PASS: Pause reason should be DOM.
PASS: Pause type should be 'node-removed'.
PASS: Pause nodeId should be expected value.
CALL STACK:
0: [F] testNodeRemovedDirect
1: [P] Global Code
Breakpoint "domNode" set to "null".
Breakpoint "domNode" set to "div#node-removed-direct-test".
-- Running test teardown.

-- Running test case: DOMBreakpoint.NodeRemoved.Direct.BreakpointDisabled
Adding "node-removed:1,HTML,1,BODY,1,DIV,0,DIV" DOM Breakpoint...
Disabling breakpoint...
Triggering breakpoint...
Breakpoint "domNode" set to "null".
Breakpoint "domNode" set to "div#node-removed-direct-test".
PASS: Should not pause for disabled breakpoint.
-- Running test teardown.

-- Running test case: DOMBreakpoint.NodeRemoved.Direct.DebuggerDisabled
Adding "node-removed:1,HTML,1,BODY,1,DIV,0,DIV" DOM Breakpoint...
Disabling all breakpoints...
Triggering breakpoint...
Breakpoint "domNode" set to "null".
Breakpoint "domNode" set to "div#node-removed-direct-test".
Enabling all breakpoints...
PASS: Should not pause when all breakpoints disabled.
-- Running test teardown.

-- Running test case: DOMBreakpoint.NodeRemoved.Direct.RemoveBreakpoint
Adding "node-removed:1,HTML,1,BODY,1,DIV,0,DIV" DOM Breakpoint...
Removing breakpoint...
Triggering breakpoint...
PASS: Should not pause for removed breakpoint.
-- Running test teardown.

-- Running test case: DOMBreakpoint.NodeRemoved.Direct.Options.Condition
Adding "node-removed:1,HTML,1,BODY,1,DIV,0,DIV" DOM Breakpoint...

Setting condition to 'false'...

Triggering breakpoint...
Breakpoint "domNode" set to "null".
Breakpoint "domNode" set to "div#node-removed-direct-test".
PASS: Should not pause.

Triggering breakpoint...
Breakpoint "domNode" set to "null".
Breakpoint "domNode" set to "div#node-removed-direct-test".
PASS: Should not pause.

Setting condition to 'true'...

Triggering breakpoint...
Breakpoint "domNode" set to "null".
Breakpoint "domNode" set to "div#node-removed-direct-test".
PASS: Should pause.

Triggering breakpoint...
Breakpoint "domNode" set to "null".
Breakpoint "domNode" set to "div#node-removed-direct-test".
PASS: Should pause.

-- Running test case: DOMBreakpoint.NodeRemoved.Direct.Options.Condition.ConsoleCommandLineAPI
Adding saved console value 'false'...
Adding "node-removed:1,HTML,1,BODY,1,DIV,0,DIV" DOM Breakpoint...

Setting condition to saved console value...

Triggering breakpoint...
Breakpoint "domNode" set to "null".
Breakpoint "domNode" set to "div#node-removed-direct-test".
PASS: Should not pause.

Triggering breakpoint...
Breakpoint "domNode" set to "null".
Breakpoint "domNode" set to "div#node-removed-direct-test".
PASS: Should not pause.

Adding saved console value 'true'...
Setting condition to saved console value...

Triggering breakpoint...
Breakpoint "domNode" set to "null".
Breakpoint "domNode" set to "div#node-removed-direct-test".
PASS: Should pause.

Triggering breakpoint...
Breakpoint "domNode" set to "null".
Breakpoint "domNode" set to "div#node-removed-direct-test".
PASS: Should pause.

-- Running test case: DOMBreakpoint.NodeRemoved.Direct.Options.Action.Log
Adding "node-removed:1,HTML,1,BODY,1,DIV,0,DIV" DOM Breakpoint...

Adding log action...

Triggering breakpoint...
Breakpoint "domNode" set to "null".
Breakpoint "domNode" set to "div#node-removed-direct-test".
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...

Triggering breakpoint...
Breakpoint "domNode" set to "null".
Breakpoint "domNode" set to "div#node-removed-direct-test".
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...
Enabling auto-continue...

Triggering breakpoint...
Breakpoint "domNode" set to "null".
Breakpoint "domNode" set to "div#node-removed-direct-test".
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing log action...

Triggering breakpoint...
Breakpoint "domNode" set to "null".
Breakpoint "domNode" set to "div#node-removed-direct-test".
PASS: Should execute breakpoint action.
PASS: Should not pause.

-- Running test case: DOMBreakpoint.NodeRemoved.Direct.Options.Actions.Evaluate
Adding "node-removed:1,HTML,1,BODY,1,DIV,0,DIV" DOM Breakpoint...

Adding evaluate action...

Triggering breakpoint...
Breakpoint "domNode" set to "null".
Breakpoint "domNode" set to "div#node-removed-direct-test".
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...

Triggering breakpoint...
Breakpoint "domNode" set to "null".
Breakpoint "domNode" set to "div#node-removed-direct-test".
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...
Enabling auto-continue...

Triggering breakpoint...
Breakpoint "domNode" set to "null".
Breakpoint "domNode" set to "div#node-removed-direct-test".
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing evaluate action...

Triggering breakpoint...
Breakpoint "domNode" set to "null".
Breakpoint "domNode" set to "div#node-removed-direct-test".
PASS: Should execute breakpoint action.
PASS: Should not pause.

-- Running test case: DOMBreakpoint.NodeRemoved.Direct.Options.Actions.Evaluate.ConsoleCommandLineAPI
Adding saved console value '1'...
Adding "node-removed:1,HTML,1,BODY,1,DIV,0,DIV" DOM Breakpoint...

Adding evaluate action using saved console value...

Triggering breakpoint...
Breakpoint "domNode" set to "null".
Breakpoint "domNode" set to "div#node-removed-direct-test".
PASS: Should execute breakpoint action.
PASS: Should pause.

Adding saved console value '2'...
Editing evaluate action using saved console value...

Triggering breakpoint...
Breakpoint "domNode" set to "null".
Breakpoint "domNode" set to "div#node-removed-direct-test".
PASS: Should execute breakpoint action.
PASS: Should pause.

Adding saved console value '3'...
Editing evaluate action using saved console value...
Enabling auto-continue...

Triggering breakpoint...
Breakpoint "domNode" set to "null".
Breakpoint "domNode" set to "div#node-removed-direct-test".
PASS: Should execute breakpoint action.
PASS: Should not pause.

Adding saved console value '4'...
Editing evaluate action using saved console value...

Triggering breakpoint...
Breakpoint "domNode" set to "null".
Breakpoint "domNode" set to "div#node-removed-direct-test".
PASS: Should execute breakpoint action.
PASS: Should not pause.
